More refactoring to remove unnecessary code
