+++ src/main/java/nl/ictu/service/v1/crypto/TokenConverter.java
@@ -1,12 +1,32 @@
 package nl.ictu.service.v1.crypto;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
-import nl.ictu.Token;
-
+import com.fasterxml.jackson.databind.ObjectMapper;
 import java.io.IOException;
+import java.io.StringWriter;
+import lombok.RequiredArgsConstructor;
+import nl.ictu.Token;
+import org.springframework.aot.hint.annotation.RegisterReflectionForBinding;
+import org.springframework.stereotype.Service;
 
-public interface TokenConverter {
-    String encode(Token token) throws IOException;
+@SuppressWarnings("DesignForExtension")
+@Service
+@RequiredArgsConstructor
+@RegisterReflectionForBinding({Token.class})
+public class TokenConverter {
 
-    Token decode(String encodedToken) throws JsonProcessingException;
+    private final ObjectMapper objectMapper;
+
+    public String encode(final Token token) throws IOException {
+
+        final StringWriter stringWriter = new StringWriter();
+        objectMapper.writeValue(stringWriter, token);
+        return stringWriter.toString();
+    }
+
+    public Token decode(final String encodedToken) throws JsonProcessingException {
+
+        return objectMapper.readValue(encodedToken, Token.class);
+    }
+
 }
